added function to refresh the popup list with new tab data when a pinned tabs is updated
