+++ popup.html
@@ -96,6 +96,6 @@
 
   <div id="priority-tabs"></div>
 
-  <script src="./scripts/popup.js"></script>
+  <script src="./scripts/popup.js" type="module"></script>
 </body>
 </html>
